+++ co-log/src/Colog.hs
@@ -1,3 +1,12 @@
+{- |
+Copyright:  (c) 2018-2019 Kowainik
+License:    MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+This package contains @mtl@ implementation of composable, contravariant and
+comonadic logging based on @co-log-core@.
+-}
+
 module Colog
        ( module Colog.Actions
        , module Colog.Core