+++ co-log/src/Colog/Contra.hs
@@ -1,7 +1,13 @@
-{-# LANGUAGE CPP #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
-{- | This module contains 'LogAction' instances of @contravariant@ classes.
+{-# LANGUAGE CPP #-}
+
+{- |
+Copyright:  (c) 2018-2019 Kowainik
+License:    MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+This module contains 'LogAction' orphan instances of @contravariant@ classes.
 -}
 
 module Colog.Contra
@@ -14,18 +20,26 @@ import Data.Functor.Contravariant (Contravariant (..))
 import Data.Functor.Contravariant.Divisible (Decidable (..), Divisible (..))
 
 import Colog.Core.Action (LogAction)
+
 import qualified Colog.Core.Action as LA
+
 
 #if !MIN_VERSION_base(4,12,0)
 instance Contravariant (LogAction m) where
     contramap = LA.cmap
-    (>$)      = (LA.>$)
+    {-# INLINE contramap#-}
+    (>$) = (LA.>$)
+    {-# INLINE (>$) #-}
 #endif
 
 instance (Applicative m) => Divisible (LogAction m) where
     divide  = LA.divide
+    {-# INLINE divide #-}
     conquer = LA.conquer
+    {-# INLINE conquer #-}
 
 instance (Applicative m) => Decidable (LogAction m) where
     lose   = LA.lose
+    {-# INLINE lose #-}
     choose = LA.choose
+    {-# INLINE choose #-}