+++ co-log/src/Colog/Monad.hs
@@ -1,5 +1,13 @@
 {-# LANGUAGE InstanceSigs #-}
 
+{- |
+Copyright:  (c) 2018-2019 Kowainik
+License:    MPL-2.0
+Maintainer: Kowainik <xrom.xkov@gmail.com>
+
+Core of the @mtl@ implementation.
+-}
+
 module Colog.Monad
        ( LoggerT (..)
        , WithLog